requirements.txt
# Core dependencies - Updated with working versions
fastapi==0.115.13
uvicorn==0.34.3
pydantic==2.11.7
pydantic-settings==2.9.1
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
asyncpg==0.30.0
aiosqlite==0.21.0
redis==6.2.0
celery==5.5.3

# AI and ML libraries
openai==1.88.0
anthropic==0.54.0
langchain==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
langsmith==0.3.45
scikit-learn==1.7.0

# Data processing
pandas==2.3.0
numpy==2.3.0
scipy==1.15.3

# Web and async
aiohttp==3.12.13
aiofiles==24.1.0
httpx==0.28.1
httpcore==1.0.9
requests==2.32.4

# Security and authentication
cryptography==45.0.4
python-jose==3.5.0
PyJWT==2.10.1
email_validator==2.2.0

# Environment and configuration
python-dotenv==1.1.0

# Testing
pytest==8.4.1
pytest-asyncio==0.24.0

# Development tools
black==25.1.0
flake8==7.2.0
mypy==1.16.1

# Monitoring and logging
structlog==25.4.0

# File handling
python-multipart==0.0.20

# Database migrations
alembic==1.16.2

# Additional utilities
click==8.2.1
tenacity==9.1.2
dnspython==2.7.0
PyYAML==6.0.2
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
autopep8==2.3.2
